Fix circular dependency fail
